Windows fullscreen window on second monitor fix
When going fullscreen on Windows, there are two calls to SetWindowPos(), the first of these is called before the window's monitor is found. getCurrentMonitor() fails to return the correct monitor index when this function is called. This results in the window going fullscreen on the monitor with index 0.

Simply removing the first call fixes the problem.